refactor(registration-react): Implement FormDistributionApi as composition of independent enhancers
