+++ applications/registration-react/src/pages/dataset-registration-page/form-distribution-api/form-configurer.jsx
@@ -1,7 +1,5 @@
 import { reduxForm } from 'redux-form';
 import _throttle from 'lodash/throttle';
-
-import { FormDistributionApiPure } from './form-distribution-api-pure';
 import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
@@ -9,6 +7,4 @@ const config = {
   asyncValidate: _throttle(asyncValidateDatasetInvokePatch, 250)
 };
 
-export const ConfiguredFormDistributionAPI = reduxForm(config)(
-  FormDistributionApiPure
-);
+export const formConfigurer = reduxForm(config);
