+++ applications/registration-react/src/pages/dataset-registration-page/form-distribution-api/resolver.jsx
@@ -1,6 +1,5 @@
 import _ from 'lodash';
 import { resolve } from 'react-resolver';
-import { ConfiguredFormDistributionAPI } from './configured-form-distribution-api';
 import { getApiByDatasetUri } from '../../../api/apis';
 
 const memoizedGetApiByDatasetUri = _.memoize(getApiByDatasetUri);
@@ -10,6 +9,4 @@ const mapProps = {
     memoizedGetApiByDatasetUri(_.get(props, 'datasetUri'), 'id,title,publisher')
 };
 
-export const ResolvedFormDistributionAPI = resolve(mapProps)(
-  ConfiguredFormDistributionAPI
-);
+export const resolver = resolve(mapProps);
